Atualização das migrations

+++ src/main/resources/db/migration/V1__create_table_consulta.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `consulta` (
  `id` bigint NOT NULL PRIMARY KEY AUTO_INCREMENT,
  `tipo` varchar(255) NOT NULL,
- `valor` double(10) NOT NULL,
+ `valor` double(10,2) NOT NULL,
  `funcionario` varchar(255) NOT NULL,
  `observacao` varchar(255) NOT NULL
 );
